requirements.txt
# After making your desired changes, update .memory/memory.md with what you did and why. Read it before each task.
dearpygui
requests
websockets

# CSI motion pipeline (backend/)
pyserial
numpy
matplotlib
